+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+### Features
+
+- Added events support in trace detail view.(https://github.com/grafana/clickhouse-datasource/pull/1128)
 
 ## 4.6.0